Add files via upload
 I'm using Cyclistic’s historical trip data to analyze and identify trends. Download the previous 12
months of Cyclistic trip data here. (Note: The datasets have a different name because Cyclistic is a
fictional company. For the purposes of this case study, the datasets are appropriate and will enable
you to answer the business questions. The data has been made available by Motivate International
Inc. under this license.) This is public data that I can use to explore how different customer types
are using Cyclistic bikes.